Add delay before running the script
DSM runs "on bootup" Task Scheduler tasks at somewhat wrong time - instead of running them after
all systemd units finished starting, the tasks are executed in the middle of the boot sequence
actually, in parallel with other systemd units. Normally the majority of critical systemd units
(disk mounts etc) have already started at that moment thanks to dependencies, but there are few
systemd units added by the installed packages which have the same boot order as esynoscheduler
service. Such units are racing with esynoscheduler, sometimes they executed before during the
boot, sometimes after. And there is at least 1 package which tries to add a synocrond task at
that time. This means that if it was able to run _after_ the script, it will overwrite changes
done by the script.

For now, add a small delay to wait for other systemd units to start before running the script.

This is a temporary solution as I want to replace it with scheduling the script execution to
post-systemd stage.